# frontend/src/features/research-universe/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandResponse": {
        "properties": {
          "commit_position": {
            "title": "Commit Position",
            "type": "integer"
          },
          "event_ids": {
            "items": {
              "type": "string"
            },
            "title": "Event Ids",
            "type": "array"
          },
          "fragment": {
            "additionalProperties": true,
            "title": "Fragment",
            "type": "object"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          }
        },
        "required": [
          "commit_position",
          "event_ids",
          "result",
          "fragment"
        ],
        "title": "CommandResponse",
        "type": "object"
      },
...
      "CorrectGapCommand": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "corrected_coverage_statement": {
            "minLength": 10,
            "title": "Corrected Coverage Statement",
            "type": "string"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "user_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Reason"
          }
        },
        "required": [
          "command_id",
          "expected_sequence",
          "corrected_coverage_statement"
        ],
        "title": "CorrectGapCommand",
        "type": "object"
      },
...
      "ExternalRef": {
        "properties": {
          "excerpt": {
            "minLength": 1,
            "title": "Excerpt",
            "type": "string"
          },
          "locator": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Locator",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "locator",
          "excerpt"
        ],
        "title": "ExternalRef",
        "type": "object"
      },
...
      "GapDecisionCommand": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "user_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Reason"
          }
        },
        "required": [
          "command_id",
          "expected_sequence"
        ],
        "title": "GapDecisionCommand",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProposeGapCommand": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "counterexample_invitation": {
            "minLength": 1,
            "title": "Counterexample Invitation",
            "type": "string"
          },
          "coverage_statement": {
            "minLength": 10,
            "title": "Coverage Statement",
            "type": "string"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "external_refs": {
            "items": {
              "$ref": "#/components/schemas/ExternalRef"
            },
            "title": "External Refs",
            "type": "array"
          },
          "matched_locators": {
            "items": {
              "type": "string"
            },
            "title": "Matched Locators",
            "type": "array"
          },
          "search_query": {
            "minLength": 1,
            "title": "Search Query",
            "type": "string"
          },
          "search_scope": {
            "default": "active",
            "enum": [
              "active",
              "legacy"
            ],
            "title": "Search Scope",
            "type": "string"
          },
          "searched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searched At"
          }
        },
        "required": [
          "command_id",
          "expected_sequence",
          "coverage_statement",
          "search_query",
          "counterexample_invitation"
        ],
        "title": "ProposeGapCommand",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Cui native test",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/gap-candidates/{candidate_id}/confirm": {
      "post": {
        "operationId": "confirm_api_v2_gap_candidates__candidate_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapDecisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm",
        "tags": [
          "research-universe-slice8-gap"
        ]
      }
    },
    "/api/v2/gap-candidates/{candidate_id}/correct": {
      "post": {
        "operationId": "correct_api_v2_gap_candidates__candidate_id__correct_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectGapCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct",
        "tags": [
          "research-universe-slice8-gap"
        ]
      }
    },
    "/api/v2/gap-candidates/{candidate_id}/reject": {
      "post": {
        "operationId": "reject_api_v2_gap_candidates__candidate_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapDecisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject",
        "tags": [
          "research-universe-slice8-gap"
        ]
      }
    },
    "/api/v2/gap-candidates/{candidate_id}/withdraw": {
      "post": {
        "operationId": "withdraw_api_v2_gap_candidates__candidate_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapDecisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw",
        "tags": [
          "research-universe-slice8-gap"
        ]
      }
    },
...
    "/api/v2/workspaces/{workspace_id}/gap-candidates": {
      "post": {
        "operationId": "propose_api_v2_workspaces__workspace_id__gap_candidates_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeGapCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose",
        "tags": [
          "research-universe-slice8-gap"
        ]
      }
    },
    "/api/v2/workspaces/{workspace_id}/landscape": {
      "get": {
        "operationId": "landscape_api_v2_workspaces__workspace_id__landscape_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Landscape",
        "tags": [
          "research-universe-slice8-gap"
        ]
      }
    },
...
  }
}